[Refactoring] split file
